backend/requirements.txt
playwright
playwright-stealth
pydantic
supabase
python-dotenv
dateparser
beautifulsoup4
duckduckgo-search
